+++ index.d.ts
@@ -9,5 +9,10 @@ declare module 'discord-logging-handler' {
         constructor(webhookURL: string, options?: DiscordLogOptions);
         constructor(webhookURL: string, level?: LogLevel)
         log(message: string, level?: LogLevel, error?: unknown): void;
+        debug(message: string, error?: unknown): void;
+        info(message: string, error?: unknown): void;
+        warning(message: string, error?: unknown): void;
+        error(message: string, error?: unknown): void;
+        critical(message: string, error?: unknown): void;
     }
 }